separate raw gff filtering tools and convert to single command
