+++ scripts/gff.py
@@ -368,231 +368,6 @@ def set_source(input_path: str, output_path: str, source: str) -> None:
     logger.info(f"Source field updated for {features.shape[0]} records")
 
 
-def filter_to_min_gene_length(
-    input_path: str, output_path: str, min_length: int
-) -> None:
-    """Filter GFF file to remove genes shorter than minimum length and their children.
-
-    Parameters
-    ----------
-    input_path : str
-        Path to input GFF file
-    output_path : str
-        Path to output GFF file
-    min_length : int
-        Minimum gene length to retain
-    """
-    logger.info(f"Filtering {input_path} to remove genes shorter than {min_length} bp")
-
-    # Read GFF file
-    features = load_gff(input_path)
-
-    original_count = features.shape[0]
-
-    # Filter to only gene features for length checking
-    genes = features[features["type"] == GffFeatureType.GENE.value]
-    logger.info(f"Found {len(genes)} gene features to check for length")
-
-    # Calculate gene lengths
-    gene_lengths = genes["end"] - genes["start"] + 1
-
-    # Identify genes to remove based on length
-    too_short_mask = gene_lengths < min_length
-    if "id" in genes.columns:
-        too_short_ids = set(genes.loc[too_short_mask, "id"].dropna())
-    else:
-        logger.warning("Column 'id' not found; cannot identify short genes by ID")
-        too_short_ids = set()
-    logger.info(f"Found {len(too_short_ids)} genes shorter than {min_length} bp")
-
-    # Remove features and their children
-    features, removed_count, indirect_count = remove_features_by_id(
-        features, too_short_ids
-    )
-
-    logger.info(f"Removing {len(too_short_ids)} genes directly (too short)")
-    logger.info(
-        f"Removing {indirect_count} features indirectly (parent gene too short)"
-    )
-    logger.info(f"Total features removed: {removed_count}")
-
-    # Write filtered GFF
-    save_gff(output_path, features)
-    logger.info(
-        f"Filter complete: {features.shape[0]}/{original_count} records retained"
-    )
-
-
-def filter_to_min_feature_length(
-    input_path: str, output_path: str, feature_types: list[str], min_length: int
-) -> None:
-    """Filter GFF file to remove small features of specified types and update gene/mRNA boundaries.
-
-    Parameters
-    ----------
-    input_path : str
-        Path to input GFF file
-    output_path : str
-        Path to output GFF file
-    feature_types : list[str]
-        List of feature types to filter by length
-    min_length : int
-        Minimum feature length to retain
-    """
-    logger.info(
-        f"Filtering {input_path} to remove {feature_types} features shorter than {min_length} bp"
-    )
-
-    # Read GFF file
-    features = load_gff(input_path)
-    original_count = features.shape[0]
-
-    # Validate feature types against schema
-    # pyrefly: ignore  # no-matching-overload
-    valid_types = set(GffFeatureType)
-    requested_types = set(feature_types)
-    invalid_types = requested_types - valid_types
-
-    if invalid_types:
-        logger.error(f"Invalid feature types requested: {sorted(invalid_types)}")
-        logger.error(f"Valid feature types: {sorted(valid_types)}")
-        raise ValueError(
-            f"Feature types {sorted(invalid_types)} are not valid GFF feature types"
-        )
-
-    logger.info(f"Validated feature types: {sorted(requested_types)} (all valid)")
-
-    # Calculate feature lengths
-    features["length"] = features["end"] - features["start"] + 1
-
-    # Find features to remove (small features of specified types)
-    small_features_mask = features["type"].isin(feature_types) & (
-        features["length"] < min_length
-    )
-    if "id" in features.columns:
-        small_feature_ids = set(features.loc[small_features_mask, "id"].dropna())
-        logger.info(f"Found {len(small_feature_ids)} small features to remove")
-    else:
-        logger.warning(
-            "Column 'id' not found; counting small features by rows instead of IDs"
-        )
-        logger.info(f"Found {int(small_features_mask.sum())} small features to remove")
-
-    # Remove small features
-    features_filtered = features[~small_features_mask]
-
-    # Group features by gene and update boundaries when hierarchical columns are available
-    genes_updated = 0
-    transcripts_updated = 0
-
-    if {"id", "parent"}.issubset(features_filtered.columns):
-        genes = features_filtered[
-            features_filtered["type"] == GffFeatureType.GENE.value
-        ]
-        mrnas = features_filtered[
-            features_filtered["type"] == GffFeatureType.MRNA.value
-        ]
-
-        # Track statistics for boundary updates
-        total_genes = len(genes)
-        total_transcripts = len(mrnas)
-
-        for _, gene in genes.iterrows():
-            gene_id = gene["id"]
-            original_gene_start = gene["start"]
-            original_gene_end = gene["end"]
-
-            # Find all features belonging to this gene (direct children and grandchildren via mRNAs)
-            gene_children = features_filtered[features_filtered["parent"] == gene_id]
-            gene_mrna_ids = gene_children[
-                gene_children["type"] == GffFeatureType.MRNA.value
-            ]["id"].tolist()
-
-            # First, update mRNA boundaries for this gene
-            for mrna_id in gene_mrna_ids:
-                mrna_features = features_filtered[
-                    features_filtered["parent"] == mrna_id
-                ]
-                if not mrna_features.empty:
-                    original_mrna = features_filtered[
-                        features_filtered["id"] == mrna_id
-                    ].iloc[0]
-                    original_mrna_start = original_mrna["start"]
-                    original_mrna_end = original_mrna["end"]
-
-                    mrna_start = mrna_features["start"].min()
-                    mrna_end = mrna_features["end"].max()
-
-                    # Update mRNA boundaries and track if changed
-                    if (
-                        mrna_start != original_mrna_start
-                        or mrna_end != original_mrna_end
-                    ):
-                        transcripts_updated += 1
-                        features_filtered.loc[
-                            features_filtered["id"] == mrna_id, "start"
-                        ] = mrna_start
-                        features_filtered.loc[
-                            features_filtered["id"] == mrna_id, "end"
-                        ] = mrna_end
-
-            # Now update gene boundaries based on updated mRNA boundaries
-            updated_gene_children = features_filtered[
-                features_filtered["parent"] == gene_id
-            ]
-            updated_mrnas = updated_gene_children[
-                updated_gene_children["type"] == GffFeatureType.MRNA.value
-            ]
-
-            if not updated_mrnas.empty:
-                new_start = updated_mrnas["start"].min()
-                new_end = updated_mrnas["end"].max()
-
-                # Update gene boundaries and track if changed
-                if new_start != original_gene_start or new_end != original_gene_end:
-                    genes_updated += 1
-                    features_filtered.loc[
-                        features_filtered["id"] == gene_id, "start"
-                    ] = new_start
-                    features_filtered.loc[features_filtered["id"] == gene_id, "end"] = (
-                        new_end
-                    )
-    else:
-        total_genes = int(
-            (features_filtered["type"] == GffFeatureType.GENE.value).sum()
-        )
-        total_transcripts = int(
-            (features_filtered["type"] == GffFeatureType.MRNA.value).sum()
-        )
-        logger.warning(
-            "Columns 'id' and/or 'parent' not found; skipping gene/mRNA boundary updates"
-        )
-
-    # Remove temporary length column
-    features_filtered = features_filtered.drop(columns=["length"])
-
-    removed_count = original_count - len(features_filtered)
-
-    # Log statistics
-    logger.info("Boundary updates:")
-    genes_percentage = (genes_updated / total_genes * 100) if total_genes > 0 else 0
-    transcripts_percentage = (
-        (transcripts_updated / total_transcripts * 100) if total_transcripts > 0 else 0
-    )
-    logger.info(
-        f"  - Genes: {genes_updated}/{total_genes} ({genes_percentage:.1f}%) had boundaries updated"
-    )
-    logger.info(
-        f"  - Transcripts: {transcripts_updated}/{total_transcripts} ({transcripts_percentage:.1f}%) had boundaries updated"
-    )
-
-    # Write filtered GFF
-    save_gff(output_path, features_filtered)
-    logger.info(
-        f"Filter complete: {removed_count} features removed, {len(features_filtered)}/{original_count} records retained"
-    )
-
-
 def filter_to_representative_transcripts(
     input_path: str, output_path: str, mode: str
 ) -> None:
@@ -733,116 +508,6 @@ def filter_to_representative_transcripts(
 
     # Remove the temporary length column
     features = features.drop(columns=["length"])
-
-    # Write filtered GFF
-    save_gff(output_path, features)
-    logger.info(
-        f"Filter complete: {features.shape[0]}/{original_count} records retained"
-    )
-
-
-def filter_to_valid_genes(
-    input_path: str, output_path: str, require_utrs: bool = True
-) -> None:
-    """Filter GFF file to remove mRNAs without required features and genes without valid mRNAs.
-
-    Parameters
-    ----------
-    input_path : str
-        Path to input GFF file
-    output_path : str
-        Path to output GFF file
-    require_utrs : bool, default True
-        If True, require mRNAs to have five_prime_UTR, CDS, and three_prime_UTR.
-        If False, only require CDS.
-    """
-    logger.info(f"Filtering {input_path} to keep only valid transcripts and genes")
-    if require_utrs:
-        logger.info(
-            "Requiring five_prime_UTR, CDS, and three_prime_UTR for valid transcripts"
-        )
-    else:
-        logger.info("Requiring only CDS for valid transcripts")
-
-    # Read GFF file
-    features = load_gff(input_path)
-    original_count = features.shape[0]
-
-    if not {"id", "parent"}.issubset(features.columns):
-        logger.warning(
-            "Columns 'id' and/or 'parent' not found; skipping valid-gene filter"
-        )
-        save_gff(output_path, features)
-        return
-
-    # Find all mRNAs
-    mrnas = features[features["type"] == GffFeatureType.MRNA.value]
-    logger.info(f"Found {len(mrnas)} mRNA features")
-
-    # Create index of parent -> children
-    features_by_parent = features.set_index("parent")
-
-    # Identify which mRNAs have the required features
-    valid_mrnas = set()
-    for mrna_id in mrnas["id"].dropna():
-        # Check if mRNA has any children
-        if mrna_id in features_by_parent.index:
-            # Get all children of this mRNA using the index
-            children = features_by_parent.loc[[mrna_id]]
-
-            # Check if it has all required feature types
-            has_cds = (children["type"] == GffFeatureType.CDS.value).any()
-
-            if require_utrs:
-                has_five_prime = (
-                    children["type"] == GffFeatureType.FIVE_PRIME_UTR.value
-                ).any()
-                has_three_prime = (
-                    children["type"] == GffFeatureType.THREE_PRIME_UTR.value
-                ).any()
-
-                if has_five_prime and has_cds and has_three_prime:
-                    valid_mrnas.add(mrna_id)
-            else:
-                if has_cds:
-                    valid_mrnas.add(mrna_id)
-
-    # Identify invalid mRNAs
-    invalid_mrnas = set(mrnas["id"].dropna()) - valid_mrnas
-    logger.info(
-        f"Found {len(valid_mrnas)} valid mRNAs and {len(invalid_mrnas)} invalid mRNAs"
-    )
-
-    # Remove invalid mRNAs and their children
-    if invalid_mrnas:
-        features, mrna_removed_count, mrna_indirect_count = remove_features_by_id(
-            features, invalid_mrnas
-        )
-        logger.info(
-            f"Removed {len(invalid_mrnas)} invalid mRNAs, {mrna_indirect_count} child features, {mrna_removed_count} total features"
-        )
-
-    # Find all genes
-    genes = features[features["type"] == GffFeatureType.GENE.value]
-    logger.info(f"Found {len(genes)} gene features")
-
-    # Identify genes with no valid mRNA children using set operations
-    all_gene_ids = set(genes["id"].dropna())
-    mrna_parent_ids = set(
-        features[(features["type"] == GffFeatureType.MRNA.value)]["parent"].dropna()
-    )
-    invalid_genes = all_gene_ids - mrna_parent_ids
-
-    logger.info(f"Found {len(invalid_genes)} genes with no valid transcripts")
-
-    # Remove invalid genes and their children
-    if invalid_genes:
-        features, gene_removed_count, gene_indirect_count = remove_features_by_id(
-            features, invalid_genes
-        )
-        logger.info(
-            f"Removed {len(invalid_genes)} invalid genes, {gene_indirect_count} child features, {gene_removed_count} total features"
-        )
 
     # Write filtered GFF
     save_gff(output_path, features)
@@ -1693,35 +1358,6 @@ def main() -> None:
     source_parser.add_argument("--output", required=True, help="Output GFF file")
     source_parser.add_argument("--source", required=True, help="Source value to set")
 
-    # Filter to minimum length command
-    length_parser = subparsers.add_parser(
-        "filter_to_min_gene_length",
-        help="Filter GFF file to remove genes shorter than minimum length and their children",
-    )
-    length_parser.add_argument("--input", required=True, help="Input GFF file")
-    length_parser.add_argument("--output", required=True, help="Output GFF file")
-    length_parser.add_argument(
-        "--min-length", required=True, type=int, help="Minimum gene length to retain"
-    )
-
-    # Filter small features command
-    small_features_parser = subparsers.add_parser(
-        "filter_to_min_feature_length",
-        help="Filter GFF file to remove small features of specified types and update gene/mRNA boundaries",
-    )
-    small_features_parser.add_argument("--input", required=True, help="Input GFF file")
-    small_features_parser.add_argument(
-        "--output", required=True, help="Output GFF file"
-    )
-    small_features_parser.add_argument(
-        "--feature-types",
-        required=True,
-        help="Comma-separated list of feature types to filter by length",
-    )
-    small_features_parser.add_argument(
-        "--min-length", required=True, type=int, help="Minimum feature length to retain"
-    )
-
     # Filter to representative transcripts command
     rep_parser = subparsers.add_parser(
         "filter_to_representative_transcripts",
@@ -1734,20 +1370,6 @@ def main() -> None:
         required=True,
         choices=["longest", "annotated", "annotated_or_longest"],
         help="Selection mode: 'longest' always uses longest transcript, 'annotated' uses canonical_transcript=1 (errors if missing), 'annotated_or_longest' tries canonical annotationfirst then falls back to longest if that annotation does not exist for even a single feature",
-    )
-
-    # Filter to valid genes command
-    valid_parser = subparsers.add_parser(
-        "filter_to_valid_genes",
-        help="Filter GFF file to remove transcripts without required features and genes without valid transcripts",
-    )
-    valid_parser.add_argument("--input", required=True, help="Input GFF file")
-    valid_parser.add_argument("--output", required=True, help="Output GFF file")
-    valid_parser.add_argument(
-        "--require-utrs",
-        choices=["yes", "no"],
-        default="yes",
-        help="Require UTRs for valid transcripts (default: yes)",
     )
 
     # Remove UTRs command
@@ -1892,16 +1514,8 @@ def main() -> None:
         filter_to_strand(args.input, args.output, args.strand)
     elif args.command == "set_source":
         set_source(args.input, args.output, args.source)
-    elif args.command == "filter_to_min_gene_length":
-        filter_to_min_gene_length(args.input, args.output, args.min_length)
-    elif args.command == "filter_to_min_feature_length":
-        filter_to_min_feature_length(
-            args.input, args.output, args.feature_types.split(","), args.min_length
-        )
     elif args.command == "filter_to_representative_transcripts":
         filter_to_representative_transcripts(args.input, args.output, args.mode)
-    elif args.command == "filter_to_valid_genes":
-        filter_to_valid_genes(args.input, args.output, args.require_utrs == "yes")
     elif args.command == "remove_exon_utrs":
         remove_exon_utrs(args.input, args.output)
     elif args.command == "compare":
